Refactor: 상세페이지 레이아웃 분리
- 게시판 별로 상세페이지 보기의 겹치는 부분을 분리(헤더, 댓글 등)
