+++ src/components/together/MarketBoardDetail.jsx
@@ -2,8 +2,7 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { MarketPostInfo } from "./MarketPostInfo";
 import CommunityPostContent from "../community/CommunityPostContent";
-import CommunityTagShareBar from "../community/CommunityTagShareBar";
-import CommunityCommentSection from "../community/CommunityCommentSection";
+import { BoardDetailLayout } from "../common/BoardDetailLayout";
 import "../../styles/together/market.css";
 
 /**
@@ -18,13 +17,6 @@ import "../../styles/together/market.css";
 export const MarketBoardDetail = ({ post }) => {
   const navigate = useNavigate();
   const [mainImage, setMainImage] = useState(post.images?.[0] || null);
-
-  const [bookmarked, setBookmarked] = useState(post.bookmarked || false);
-  const [bookmarkCount, setBookmarkCount] = useState(post.bookmarkCount || 0);
-  const handleBookmarkClick = () => {
-    setBookmarked((prev) => !prev);
-    setBookmarkCount((prev) => prev + (bookmarked ? -1 : 1));
-  };
 
   const handleEdit = () => {
     navigate(`/together/${post.category}/write`, {
@@ -41,9 +33,8 @@ export const MarketBoardDetail = ({ post }) => {
   };
 
   return (
-    <div className="community-detail-container">
+    <BoardDetailLayout post={post}>
       <MarketPostInfo post={post} onEdit={handleEdit} onDelete={handleDelete} />
-
       <div className="row g-5 mt-3">
         <div className="col-md-5">
           {post.images && post.images.length > 0 && (
@@ -75,15 +66,6 @@ export const MarketBoardDetail = ({ post }) => {
           <CommunityPostContent post={post} stripImages={true} />
         </div>
       </div>
-
-      <CommunityTagShareBar
-        tags={post.tags}
-        likes={post.likes}
-        bookmarked={bookmarked}
-        bookmarkCount={bookmarkCount}
-        onBookmarkToggle={handleBookmarkClick}
-      />
-      <CommunityCommentSection postId={post.id} commentList={post.comments} />
-    </div>
+    </BoardDetailLayout>
   );
 };
